requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
idna==2.9
lxml==4.5.0
mysqlclient==1.4.6
Pillow==7.0.0
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
